Initial implementation
Fs requires only two methods, the first one to read the file as a binary vector, and the second one to write a binary slice to a file.
Binary to UTF-8 conversion can be auto-generated, but invalid chars will be replaced.

Only the wasm version is implemented as of right now.